requirements.txt
pandas
numpy
matplotlib
seaborn
nltk
wordcloud
scikit-learn
imbalanced-learn
apache-airflow
apache-airflow
